Put grading criterions together
